refactor: Navbar를  페이지에 직접 작성하는 방식으로 수정
