+++ src/pages/FormShare.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from 'react'
-import { useParams } from 'react-router-dom'
+import { Link, useParams } from 'react-router-dom'
+import { css } from '@emotion/react'
 import FormEnter from '@/components/form/vote/FormEnter'
 import FormInfo from '@/components/form/vote/FormInfo'
 import SelectTrack from '@/components/form/vote/SelectTrack'
@@ -9,7 +10,8 @@ import { useGetFormData } from '@/hooks/useGetFormData'
 import { FormDataFromUser } from '@/models/form'
 import useNavbar from '@/hooks/useNavbar'
 import ShareLink from '@/components/form/share/ShareLink'
-import BackNavbar from '@/components/form/BackNavbar'
+import { Button } from '@/components/shared/button'
+import { ReactComponent as BackIcon } from '@/assets/icons/back.svg'
 
 export default function FormSharePage() {
   const [step, setStep] = useState<'시작' | '정보' | '투표' | '완료'>('시작')
@@ -25,7 +27,13 @@ export default function FormSharePage() {
 
   useEffect(() => {
     updateNavbar({
-      left: <BackNavbar />,
+      left: (
+        <div css={navbarStyles}>
+          <Link to={`/myforms`}>
+            <Button variant={'secondary'} size={'icon'} icon={<BackIcon />} />
+          </Link>
+        </div>
+      ),
       right: <ShareLink />,
     })
   }, [updateNavbar])
@@ -55,3 +63,11 @@ export default function FormSharePage() {
     </PageLayout>
   )
 }
+
+const navbarStyles = css`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 4px;
+  height: 100%;
+`
