+++ src/pages/Home.tsx
@@ -1,4 +1,6 @@
 import { useEffect } from 'react'
+import { Link } from 'react-router-dom'
+import { css } from '@emotion/react'
 import PageLayout from '@/components/shared/PageLayout'
 import useNavbar from '@/hooks/useNavbar'
 import SignOut from '@/components/shared/SignOut'
@@ -6,7 +8,7 @@ import { useGetAllFormData } from '@/hooks/useGetAllFormData'
 import AllFormList from '@/components/home/AllFormList'
 import Introduction from '@/components/home/Introduction'
 import { useUser } from '@/hooks/useUser'
-import Navbar from '@/components/home/Navbar'
+import { Text } from '@/components/shared/text'
 
 export default function HomePage() {
   const { updateNavbar } = useNavbar()
@@ -15,7 +17,17 @@ export default function HomePage() {
 
   useEffect(() => {
     updateNavbar({
-      left: <Navbar />,
+      left: (
+        <Link to={`/`}>
+          <Text variant={'heading1'}>
+            <img
+              css={imageStyles}
+              src={'/images/logo.png'}
+              alt="go-to-track 로고"
+            />
+          </Text>
+        </Link>
+      ),
       right: user && <SignOut />,
     })
   }, [updateNavbar, user])
@@ -27,3 +39,7 @@ export default function HomePage() {
     </PageLayout>
   )
 }
+
+const imageStyles = css`
+  height: 44px;
+`
