+++ src/pages/MyForms.tsx
@@ -1,12 +1,14 @@
 import { useEffect } from 'react'
 import { Link } from 'react-router-dom'
+import { css } from '@emotion/react'
 import PageLayout from '@/components/shared/PageLayout'
 import { Button } from '@/components/shared/button'
 import useNavbar from '@/hooks/useNavbar'
 import FormList from '@/components/myForms/FormList'
 import { useGetUserFormList } from '@/hooks/useGetUserFormList'
 import SignOut from '@/components/shared/SignOut'
-import Navbar from '@/components/myForms/Navbar'
+import { Text } from '@/components/shared/text'
+import { ReactComponent as LogoIcon } from '@/assets/icons/logo.svg'
 
 export default function MyFormsPage() {
   const { data } = useGetUserFormList()
@@ -15,7 +17,20 @@ export default function MyFormsPage() {
 
   useEffect(() => {
     updateNavbar({
-      left: <Navbar />,
+      left: (
+        <div css={navbarStyles}>
+          <Link to={`/`} aria-label={'홈으로'}>
+            <Text variant={'heading1'}>
+              <Button variant={'secondary'} size={'icon'} icon={<LogoIcon />} />
+            </Text>
+          </Link>
+          <div css={myformsWrapperStyles}>
+            <Text css={myformsStyles} variant={'heading2'}>
+              My forms
+            </Text>
+          </div>
+        </div>
+      ),
       right: <SignOut />,
     })
   }, [updateNavbar])
@@ -29,3 +44,21 @@ export default function MyFormsPage() {
     </PageLayout>
   )
 }
+
+const navbarStyles = css`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 4px;
+  height: 100%;
+`
+
+const myformsWrapperStyles = css`
+  height: 44px;
+  padding: 12px 0;
+`
+
+const myformsStyles = css`
+  font-size: 16px;
+  line-height: 20px;
+`
