+++ src/pages/Vote.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from 'react'
-import { useNavigate, useParams } from 'react-router-dom'
+import { Link, useNavigate, useParams } from 'react-router-dom'
+import { css } from '@emotion/react'
 import FormEnter from '@/components/form/vote/FormEnter'
 import FormInfo from '@/components/form/vote/FormInfo'
 import SelectTrack from '@/components/form/vote/SelectTrack'
@@ -10,7 +11,7 @@ import { FormDataFromUser } from '@/models/form'
 import { useSaveVoteData } from '@/hooks/useSaveVoteData'
 import { VoteData } from '@/models/vote'
 import useNavbar from '@/hooks/useNavbar'
-import Navbar from '@/components/form/vote/Navbar'
+import { Text } from '@/components/shared/text'
 
 export default function VotePage() {
   const [step, setStep] = useState<'시작' | '정보' | '투표' | '완료'>('시작')
@@ -44,7 +45,17 @@ export default function VotePage() {
 
   useEffect(() => {
     updateNavbar({
-      left: <Navbar />,
+      left: (
+        <Link to={`/`}>
+          <Text variant={'heading1'}>
+            <img
+              css={imageStyles}
+              src={'/images/logo.png'}
+              alt="go-to-track 로고"
+            />
+          </Text>
+        </Link>
+      ),
       right: null,
     })
   }, [updateNavbar])
@@ -92,3 +103,7 @@ export default function VotePage() {
     </PageLayout>
   )
 }
+
+const imageStyles = css`
+  height: 44px;
+`
